Force plain (not gzipped) aws response on aiohttp request
When using aiobotocore with dynamodb, requests fail on crc32
checksum computation as soon as the response data reaches ~5KB.

It seems that this is  due to aws sending data back in gzip
compressed format, aiohttp automatically uncompressing it
and crc32 checksum being computed on the uncompressed data bytes.
While crc32 has to be computed on compressed data in order
to pass.

The same requests work as expected when using botocore, this
commit forces aws not to use gzip compression.