Copy output protocol test from upstream botocore
The output procotol test cover the code in parsers.py
They test supported output protocols and eventstreams.

The `test_protocols.py` file was copied, cleared of any code specific to
input testing, then adjusted to work with asyncio. The most effort was
on the `FakeStreamReader` class that is inspired by the eventstreams
tests, though needed to be slightly different to get the tests to pass.