Create common interface for aio response classes
Having the `parse` method async on all parsers means
the need for checking that `parse` is a coroutine is not needed
anymore. The check for coroutine or not moved into the parser where
a private method can be async or sync. This cleans up the public
interface of the aio response classes.